Create new poll flow

+++ oops-all-ees-lab3/src/pages/poll-list.js
@@ -6,6 +6,7 @@ import getFirebase from '../../firebase'
 import NewPollModal from '../components/new-poll-modal'
 import {isloggedin} from '../components/auth'
 import Layout from '../components/layout'
+import swal from 'sweetalert'
 
 
 class PollListPage extends React.Component {
@@ -13,45 +14,27 @@ class PollListPage extends React.Component {
         super(props);
         this.state = {polls:[],modalShow:false}
 
-        this.newPoll = this.newPoll.bind(this);
         this.deletePoll = this.deletePoll.bind(this);
         this.componentDidMount = this.componentDidMount.bind(this)
         this.updateList = this.updateList.bind(this)
         this.setModalShow =  this.setModalShow.bind(this)
     }
 
-    newPoll () {
-        const lazyApp = import('firebase/app')
-        const lazyDatabase = import('firebase/database')
+    deletePoll(deleteID, deleteTitle) {
+        swal("Delete Poll?", "Are you sure you want to delete \"" + deleteTitle + "\"", "warning", {buttons: ["Cancel", "Delete"], dangerMode:true})
+        .then((value) => {
+            if(value) {
+                const lazyApp = import('firebase/app')
+                const lazyDatabase = import('firebase/database')
 
-        Promise.all([lazyApp, lazyDatabase]).then(([f, fdb]) => {
-            const database = fdb.getDatabase(getFirebase(f))
-
-            const newPollKey = fdb.push(fdb.child(fdb.ref(database), 'polls')).key
-            const pollData = {
-                id: newPollKey,
-                title: "Poll " + newPollKey,
-                author: "me"
+                Promise.all([lazyApp, lazyDatabase]).then(([f, fdb]) => {
+                    const database = fdb.getDatabase(getFirebase(f))
+                    const dataref = fdb.ref(database, '/polls/' + deleteID)
+                    return fdb.remove(dataref)
+                })
+                this.updateList()
             }
-
-            const updates = {}
-            updates['/polls/' + newPollKey] = pollData
-            
-            return fdb.update(fdb.ref(database), updates)
         })
-        this.updateList()
-    }
-
-    deletePoll(deleteID) {
-        const lazyApp = import('firebase/app')
-        const lazyDatabase = import('firebase/database')
-
-        Promise.all([lazyApp, lazyDatabase]).then(([f, fdb]) => {
-            const database = fdb.getDatabase(getFirebase(f))
-            const dataref = fdb.ref(database, '/polls/' + deleteID)
-            return fdb.remove(dataref)
-        })
-        this.updateList()
     }
 
     componentDidMount() {
@@ -92,15 +75,13 @@ class PollListPage extends React.Component {
         return (
             <Layout pageTitle="Poll List: Oops All EEs Doodle">
             <main>
-
-                <title>Oops All EEs Lab 3</title>
                 <div style={{backgroundColor:"#ffffff", borderRadius:"10px", border:"1px solid #000000", width:"90%", margin:"0 auto"}}>
                     <table className="table table-striped">
                         <thead>
                             <tr>
-                                <th style={{verticalAlign:'middle'}}>Poll</th>
-                                <th style={{verticalAlign:'middle'}}>Author</th>
-                                <th style={{verticalAlign:'middle', padding:'auto 0'}}>
+                                <th style={{verticalAlign:'middle', width:"40%"}}>Poll</th>
+                                <th style={{verticalAlign:'middle', width:"40%"}}>Author</th>
+                                <th style={{verticalAlign:'middle', width:"20%", padding:'auto 0'}}>
                                     <div className="row">
                                         <p style={{verticalAlign:'middle',width:'80px',margin:"auto 0"}}>Actions</p>
                                         {
